requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
httpx>=0.27.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
Pillow>=10.0.0
peft>=0.10.0
python-multipart>=0.0.6
